requirements.txt
asgiref==3.8.1
Django==5.1.5
django-sendgrid-v5==1.2.4
djangorestframework==3.15.2
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.1
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tzdata==2025.1
